refactor(workspace): extract useLayoutLoader hook from WorkspaceContext
